+++ neo_finrl/yahoo_finance/stocktrading_env.py
@@ -34,7 +34,7 @@ class StockTradingEnv(gym.Env):
         self.initial_account__reset = self.initial_account
         self.account = self.initial_account__reset
         self.day_npy = self.ary[self.day]
-        self.stocks = np.zeros(self.stock_dim, dtype=np.float32)  # multi-stack
+        self.stocks = np.zeros(self.stock_dim, dtype=np.float64)  # multi-stack
 
         self.total_asset = self.account + (self.day_npy[:self.stock_dim] * self.stocks).sum()
         self.episode_return = 0.0  # Compatibility for ElegantRL 2020-12-21
@@ -51,7 +51,7 @@ class StockTradingEnv(gym.Env):
     def reset(self):
         self.initial_account__reset = self.initial_account * rd.uniform(0.9, 1.1)  # reset()
         self.account = self.initial_account__reset
-        self.stocks = np.zeros(self.stock_dim, dtype=np.float32)
+        self.stocks = np.zeros(self.stock_dim, dtype=np.float64)
         self.total_asset = self.account + (self.day_npy[:self.stock_dim] * self.stocks).sum()
         # total_asset = account + (adjcp * stocks).sum()
 
@@ -61,7 +61,7 @@ class StockTradingEnv(gym.Env):
 
         state = np.hstack((self.account * 2 ** -16,
                            self.day_npy * 2 ** -8,
-                           self.stocks * 2 ** -12,), ).astype(np.float32)
+                           self.stocks * 2 ** -12,), ).astype(np.float64)
         return state
 
     def step(self, action):
@@ -88,7 +88,7 @@ class StockTradingEnv(gym.Env):
 
         state = np.hstack((self.account * 2 ** -16,
                            self.day_npy * 2 ** -8,
-                           self.stocks * 2 ** -12,), ).astype(np.float32)
+                           self.stocks * 2 ** -12,), ).astype(np.float64)
 
         next_total_asset = self.account + (self.day_npy[:self.stock_dim] * self.stocks).sum()
         reward = (next_total_asset - self.total_asset) * 2 ** -16  # notice scaling!
@@ -108,7 +108,7 @@ class StockTradingEnv(gym.Env):
     def load_training_data_for_multi_stock(if_load=True):  # need more independent
         npy_path = './FinanceMultiStock.npy'
         if if_load and os.path.exists(npy_path):
-            data_ary = np.load(npy_path).astype(np.float32)
+            data_ary = np.load(npy_path).astype(np.float64)
             assert data_ary.shape[1] == 5 * 30
             return data_ary
         else:
